+++ packages/backend/src/ab/errors.ts
@@ -1,9 +1,3 @@
-export class ExperimentAlreadyExists extends Error {
-  constructor() {
-    super('Experiment already exists');
-  }
-}
-
 export class VariantAlreadyExists extends Error {
   constructor() {
     super('Variant already exists');
