+++ packages/backend/src/environments/environments.module.ts
@@ -2,10 +2,9 @@ import { Module } from '@nestjs/common';
 import { PrismaService } from '../prisma.service';
 import { EnvironmentsService } from './environments.service';
 import { EnvironmentsController } from './environments.controller';
-import { AbModule } from '../ab/ab.module';
 
 @Module({
-  imports: [AbModule],
+  imports: [],
   controllers: [EnvironmentsController],
   providers: [PrismaService, EnvironmentsService],
   exports: [EnvironmentsService],
