+++ packages/backend/src/environments/environments.service.ts
@@ -3,7 +3,7 @@ import camelcase from 'camelcase';
 import { ExperimentStatus } from 'src/ab/types';
 import { FlagStatus } from '../flags/flags.status';
 import { PrismaService } from '../prisma.service';
-import { FlagAlreadyExists } from './errors';
+import { FlagAlreadyExists, ExperimentAlreadyExists } from './errors';
 
 @Injectable()
 export class EnvironmentsService {
@@ -266,5 +266,66 @@ export class EnvironmentsService {
     }
 
     return roles.includes(environmentOfProject.role);
+  }
+
+  experimentsByEnv(envId: string) {
+    return this.prisma.experiment.findMany({
+      orderBy: {
+        createdAt: 'desc',
+      },
+      where: {
+        ExperimentEnvironment: {
+          some: {
+            environmentId: envId,
+          },
+        },
+      },
+    });
+  }
+
+  async createExperimentEnv(envId: string, name: string, description: string) {
+    const experimentKey = camelcase(name);
+
+    const existingExperiment =
+      await this.prisma.experimentEnvironment.findFirst({
+        where: {
+          environmentId: envId,
+          experiment: {
+            key: experimentKey,
+          },
+        },
+      });
+
+    if (existingExperiment) {
+      throw new ExperimentAlreadyExists();
+    }
+
+    const experiment = await this.prisma.experiment.create({
+      data: {
+        name,
+        description,
+        key: experimentKey,
+      },
+    });
+
+    const variationName = `${experiment.name} Control`;
+    await this.prisma.variant.create({
+      data: {
+        key: camelcase(variationName),
+        name: variationName,
+        isControl: true,
+        experimentUuid: experiment.uuid,
+        description: `This is the control variant of the ${experiment.name} experiment`,
+      },
+    });
+
+    await this.prisma.experimentEnvironment.create({
+      data: {
+        experimentId: experiment.uuid,
+        environmentId: envId,
+      },
+    });
+
+    return experiment;
   }
 }
