+++ packages/backend/src/environments/errors.ts
@@ -3,3 +3,9 @@ export class FlagAlreadyExists extends Error {
     super('Flag already exists');
   }
 }
+
+export class ExperimentAlreadyExists extends Error {
+  constructor() {
+    super('Experiment already exists');
+  }
+}
